+++ src/main/java/btrplace/plan/action/KillVM.java
@@ -18,78 +18,47 @@
 
 package btrplace.plan.action;
 
-import btrplace.model.Mapping;
 import btrplace.model.Model;
 import btrplace.plan.Action;
 
 import java.util.UUID;
 
 /**
- * Prepare a VM for being deployed.
+ * An action to destroy a VM that can be in any state.
  *
  * @author Fabien Hermenier
  */
-public class InstantiateVM extends Action {
+public class KillVM extends Action {
 
     private UUID id;
+
+    private UUID host;
 
     /**
      * Make a new action.
      *
-     * @param vm the VM to instantiate.
+     * @param vm   the VM to kill
+     * @param host its location if any, {@code null} otherwise
+     * @param st   the moment the action starts
+     * @param ed   the moment the action ends
      */
-    public InstantiateVM(UUID vm, int st, int ed) {
+    public KillVM(UUID vm, UUID host, int st, int ed) {
         super(st, ed);
-        this.id = vm;
+        id = vm;
+        this.host = host;
     }
 
     /**
-     * Put the VM in the waiting state if it does not already belong
-     * to the mapping.
+     * Get the VM location.
      *
-     * @param m the model to modify
-     * @return {@code true} iff successful
+     * @return the node identifier if the VM is hosted somewhere. Otherwise, {@code null}
      */
-    @Override
-    public boolean apply(Model m) {
-        Mapping map = m.getMapping();
-
-        if (!map.getAllVMs().contains(id)) {
-            map.addWaitingVM(id);
-            return true;
-        }
-        return false;
+    public UUID getNode() {
+        return host;
     }
 
     /**
-     * Test if this action is equals to another object.
-     *
-     * @param o the object to compare with
-     * @return true if ref is an instance of Instantiate and if both
-     *         instance involve the same virtual machine
-     */
-    @Override
-    public boolean equals(Object o) {
-        if (o == null) {
-            return false;
-        } else if (o == this) {
-            return true;
-        } else if (o.getClass() == this.getClass()) {
-            InstantiateVM that = (InstantiateVM) o;
-            return this.id.equals(that.id);
-        }
-        return false;
-    }
-
-    @Override
-    public int hashCode() {
-        int res = getEnd();
-        res = getStart() + 31 * res;
-        return id.hashCode() + 31 * res;
-    }
-
-    /**
-     * Get the VM to instantiate.
+     * Get the VM to kill.
      *
      * @return the VM identifier
      */
@@ -98,7 +67,31 @@ public class InstantiateVM extends Action {
     }
 
     @Override
-    public String toString() {
-        return new StringBuilder("instantiate(vm=").append(id).append(')').toString();
+    public boolean apply(Model i) {
+        return i.getMapping().removeVM(id);
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (o == null) {
+            return false;
+        } else if (o == this) {
+            return true;
+        } else if (o.getClass() == this.getClass()) {
+            KillVM that = (KillVM) o;
+            return this.id.equals(that.id) &&
+                    ((host == null && that.host == null) || (host != null && host.equals(that.host))) &&
+                    this.getStart() == that.getStart() &&
+                    this.getEnd() == that.getEnd();
+        }
+        return false;
+    }
+
+    @Override
+    public int hashCode() {
+        int res = getEnd();
+        res = getStart() + 31 * res;
+        res = res * 31 + (host != null ? host.hashCode() : 0);
+        return id.hashCode() + 31 * res;
     }
 }